+++ libbs/decompilers/angr/interface.py
@@ -1,5 +1,6 @@
 import logging
 import os
+from collections import defaultdict
 from typing import Optional, Dict, List
 from pathlib import Path
 
@@ -480,3 +481,105 @@ class AngrInterface(DecompilerInterface):
 
         return func_addr
 
+    @staticmethod
+    def line_map_from_decompilation(dec):
+        import ailment
+        from angr.analyses.decompiler.structured_codegen.c import CStructuredCodeWalker, CFunctionCall, CIfElse, CIfBreak
+
+        if dec is None or dec.codegen is None:
+            return None
+
+        codegen = dec.codegen
+        base_addr = dec.project.loader.main_object.image_base_delta
+        if hasattr(dec, "unoptimized_ail_graph"):
+            nodes = dec.unoptimized_ail_graph.nodes
+        else:
+            l.critical(f"You are likely using an older version of angr that has no unoptimized_ail_graph."
+                       f" Using clinic_graph instead, results will be less accurate...")
+            nodes = dec.clinic.cc_graph.nodes
+
+        # get the mapping of the original AIL graph
+        mapping = defaultdict(set)
+        ail_node_addr_map = {
+            node.addr: node for node in nodes
+        }
+        for addr, ail_block in ail_node_addr_map.items():
+            # get instructions of this block
+            try:
+                vex_block = dec.project.factory.block(addr)
+            except Exception:
+                continue
+
+            ail_block_stmts = [stmt for stmt in ail_block.statements if not isinstance(stmt, ailment.statement.Label)]
+            if not ail_block_stmts:
+                continue
+
+            next_ail_stmt_idx = 0
+            for ins_addr in vex_block.instruction_addrs:
+                next_ail_stmt_addr = ail_block_stmts[next_ail_stmt_idx].ins_addr
+                mapping[next_ail_stmt_addr].add(ins_addr)
+                if ins_addr == next_ail_stmt_addr:
+                    next_ail_stmt_idx += 1
+                if next_ail_stmt_idx >= len(ail_block_stmts):
+                    break
+
+        # node to addr map
+        ailaddr_to_addr = defaultdict(set)
+        for k, v in mapping.items():
+            for v_ in v:
+                ailaddr_to_addr[k - base_addr].add(v_ - base_addr)
+
+        codegen.show_externs = False
+        codegen.regenerate_text()
+
+        decompilation = codegen.text
+        if not decompilation:
+            return
+
+        try:
+            first_code_pos = codegen.map_pos_to_addr.items()[0][0]
+        except Exception:
+            return
+
+        # map the position start to an address
+        pos_addr_map = defaultdict(set)
+        for start, pos_map in codegen.map_pos_to_addr.items():
+            obj = pos_map.obj
+            if not hasattr(obj, "tags"):
+                continue
+
+            # leads to mapping at the beginning of loops, so skip.
+            # see kill.o binary for send_signals
+            if isinstance(obj, CIfElse):
+                continue
+
+            ins_addr = obj.tags.get("ins_addr", None)
+            if ins_addr:
+                pos_addr_map[start].add(ins_addr - base_addr)
+
+        # find every line
+        line_end_pos = [i for i, x in enumerate(decompilation) if x == "\n"]
+        line_to_addr = defaultdict(set)
+        last_pos = len(decompilation) - 1
+        line_to_addr[str(1)].add(codegen.cfunc.addr - base_addr)
+        for i, pos in enumerate(line_end_pos[:-1]):
+            if pos == last_pos:
+                break
+
+            curr_end = line_end_pos[i+1] - 1
+            # check if this is the variable decs and header
+            if curr_end < first_code_pos:
+                line_to_addr[str(i+2)].add(codegen.cfunc.addr - base_addr)
+                continue
+
+            # not header, real code
+            for p_idx in range(pos+1, curr_end+1):
+                if p_idx in pos_addr_map:
+                    # line_to_addr[str(i+1)].update(pos_addr_map[p_idx])
+                    for ail_ins_addr in pos_addr_map[p_idx]:
+                        if ail_ins_addr in ailaddr_to_addr:
+                            line_to_addr[str(i+2)].update(ailaddr_to_addr[ail_ins_addr])
+                        else:
+                            line_to_addr[str(i+2)].add(ail_ins_addr)
+
+        return line_to_addr
